+++ src/imp/fallback/imp.rs
@@ -130,6 +130,7 @@ macro_rules! atomic {
         unsafe impl Sync for $atomic_type {}
 
         no_fetch_ops_impl!($atomic_type, $int_type);
+        bit_opts_fetch_impl!($atomic_type, $int_type);
         impl $atomic_type {
             #[inline]
             pub(crate) const fn new(v: $int_type) -> Self {