+++ src/imp/interrupt/mod.rs
@@ -515,6 +515,7 @@ macro_rules! atomic_int {
 
         #[cfg(not(all(target_arch = "msp430", not(feature = "critical-section"))))]
         no_fetch_ops_impl!($atomic_type, $int_type);
+        bit_opts_fetch_impl!($atomic_type, $int_type);
         #[cfg(not(all(target_arch = "msp430", not(feature = "critical-section"))))]
         impl $atomic_type {
             #[inline]
@@ -572,6 +573,7 @@ macro_rules! atomic_int {
         atomic_int!(base, $atomic_type, $int_type, $align);
         atomic_int!(cas, $atomic_type, $int_type);
         no_fetch_ops_impl!($atomic_type, $int_type);
+        bit_opts_fetch_impl!($atomic_type, $int_type);
         impl $atomic_type {
             #[inline]
             #[cfg_attr(all(debug_assertions, not(portable_atomic_no_track_caller)), track_caller)]